+++ src/renderer/settings_page/components/ShortcutRecorder.js
@@ -1,13 +1,19 @@
 import React, { useState, useEffect, useRef } from 'react';
 import { debounce } from 'lodash';
+import {
+  MdKeyboardCommandKey,
+  MdOutlineKeyboardControlKey,
+  MdOutlineKeyboardOptionKey,
+} from "react-icons/md";
+import { ImShift } from "react-icons/im";
 
 const HotkeyPill = ({ accelerator }) => {
   const humanizedKey = (key) => {
     const keyMap = {
-      'Meta': '⌘',
-      'Control': '⌃',
-      'Alt': '⌥',
-      'Shift': '⇧',
+      'Meta': <MdKeyboardCommandKey />,
+      'Control': <MdOutlineKeyboardControlKey />,
+      'Alt': <MdOutlineKeyboardOptionKey />,
+      'Shift': <ImShift />,
     };
 
     return keyMap[key] || key;
@@ -25,7 +31,6 @@ const HotkeyPill = ({ accelerator }) => {
 };
 
 const ShortcutRecorder = ({ accelerator, onCheck, onChange }) => {
-  console.log('Rendering ShortcutRecorder with accelerator:', accelerator);
   const isMac = window.electronAPI.isMac;
 
   const [isRecording, setIsRecording] = useState(false);
@@ -46,11 +51,9 @@ const ShortcutRecorder = ({ accelerator, onCheck, onChange }) => {
       Equal: '=',
       BracketLeft: '[',
       BracketRight: ']',
-      Backslash: '\\',
       Semicolon: ';',
       Quote: "'",
       Backquote: '`',
-      Comma: ',',
       Period: '.',
       Slash: '/',
     };
@@ -114,15 +117,11 @@ const ShortcutRecorder = ({ accelerator, onCheck, onChange }) => {
   };
 
   const stopRecording = async (keys) => {
-    console.log('Stopping recording with keys:', keys);
-
     if (validateShortcut(keys)) {
       const electronKey = keys.join('+')
 
       if (await onCheck(electronKey)) {
         onChange(electronKey)
-      } else {
-        console.log('Shortcut cannot be registered:', electronKey);
       }
     }
 
@@ -148,15 +147,14 @@ const ShortcutRecorder = ({ accelerator, onCheck, onChange }) => {
     } else {
       if (event.ctrlKey) modifiers.push('Control');
     }
+
     if (event.altKey) modifiers.push('Alt');
     if (event.shiftKey) modifiers.push('Shift');
 
-    const allKeys = [...modifiers];
-
     const mainKey = codeToKey(event.code);
-    if (mainKey) allKeys.push(mainKey);
+    if (mainKey) modifiers.push(mainKey);
 
-    setCurrentKeys(allKeys);
+    setCurrentKeys(modifiers);
   };
 
   const handleKeyUp = (event) => {